+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next Release
 * FEATURE: Ability to use just total and update without passing an iterable
 * FIX: Multi-threaded progress bars now finish correctly
+* FIX: Multiline postfix now correctly works together with ANSI escape codes (e.g. for colorful output)
 
 ## 1.4.1 (2022-03-10)
 * FIX: Progress bars on 32bit systems use the correct integer type now (`UInt` -> `UInt64`)